Move 'schema_registry' fixture to the top-level conftest.py
The same fixture will be needed for both source data validation and
build tool tests.
