+++ tests/__init__.py
@@ -1,4 +1,7 @@
 from pathlib import Path
 
 _FOLDER = Path(__file__).absolute().parent
+
 PROJECT_ROOT = _FOLDER.parent
+
+SCHEMAS = PROJECT_ROOT / "schemas"
